adding flags and dimension reduction
This version creates a function for dimension reduction, adds flags for specific label id categories that EDA showed as potentially predictive for certain groups.  I attempted to fix the label id join issue causing a majority of the observations to have NA for label id and subsequently many other later variables, but it is still running so I don't know if it works.